+++ Programa 9.cpp
@@ -7,8 +7,8 @@
 #include <string> //Permite el uso de cadenas de texto en nuestro programa.
 
 using namespace std; //Lo usamos para importar todas las funciones de la bibilioteca stl y, así, evitar introducir muchas funciones empezando por "std::".
-int temperaturaC;
-int temperaturaF;
+float temperaturaC;
+float temperaturaF;
 int decision;
 
 int main (){
